Clean up file-level side effects, misplaced namespace, and duplicated date parsing
#373: Move chubes_ai_library_error hook registration from ConversationManager.php
(file-level side effect in autoloaded class) to datamachine_register_core_actions()
in DataMachineActions.php where all action registrations live.

#372: Move SiteContext.php from Engine/AI/Directives to Core/WordPress namespace.
It is a data provider for cached WP metadata, not a directive. Remove file-level
add_action('init') side effect, register cache invalidation hook in bootstrap.php.
Update SiteContextDirective.php import.

#371: Add canonical DailyMemory::parse_date() static method. Replace duplicate
implementations in DailyMemoryAbilities::parseDate() and MemoryCommand::parse_date()
with calls to the shared method.

Closes #373, closes #372, closes #371
