+++ inc/Engine/AI/ConversationManager.php
@@ -338,21 +338,3 @@ class ConversationManager {
 		return self::formatToolResultMessage( $tool_name, $tool_result, array(), false, $turn_count );
 	}
 }
-
-// AI library error logging - universal handler for all AI interactions (pipeline agents, chat agents)
-add_action(
-	'chubes_ai_library_error',
-	function ( $error_data ) {
-		do_action(
-			'datamachine_log',
-			'error',
-			'AI Library Error: ' . $error_data['component'] . ' - ' . $error_data['message'],
-			array(
-				'component' => $error_data['component'],
-				'message'   => $error_data['message'],
-				'context'   => $error_data['context'],
-				'timestamp' => $error_data['timestamp'],
-			)
-		);
-	}
-);
